+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix shader compilation error on Nvidia GPUs
 - Fix cubemap textures on Nvidia GPUs
 - Fix truncation of open orbits at escape
+- Truncate escaping closed orbit
 
 # v0.5.1 - 2019-08-28
 - Fix crash on Windows due to uninitialized data